Added Bolts::Scope::Transient, ::Prototype sanity
After some thinking, I realized that binding to the Proxy returned during
resolution would result in some potentially very weird behavior when using what
was prototype scope. I've renamed the original prototype scope as transient and
added a new scope named prototype, which does a very small amount of caching
using a callframe trick to grab a variable from the outer lexical scope and use
that for caching. I think it works... the tests work at least. I probably need
to test it more thoroughly to be sure I got it right.

And added most of the rest of the docs